Refactor from merge with #1099, make template list stateful
